% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_bind.R
\name{force_bind}
\alias{force_bind}
\title{Force rbind of two data frames, even though colnames do not match}
\usage{
force_bind(df1, df2)
}
\description{
This is a function to force rbind of two data frames. Use with caution.
}
\examples{
force_bind()
}
\keyword{rbind}
