% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_frame.R
\name{results_frame}
\alias{results_frame}
\title{Make data.frame from regression models for easy export.}
\usage{
results_frame(..., standard_errors = NULL, t.values = FALSE,
  digits = 2, omit.coef = NULL, omit.coef.label = NULL,
  omit.stat = c("R-squared", "DF"), stars = TRUE,
  intercept.placement = "bottom", covariate.labels = NULL,
  intercept.label = "Constant")
}
\description{
This is a function to create data frame out of regression results.
}
\examples{
results_frame()
}
\keyword{export}
\keyword{tables}
